requirements.txt
streamlit
streamlit-authenticator
xgboost
scikit-learn
pandas
pyyaml
joblib
shap

